requirements.txt
fastapi
uvicorn
requests
PyPDF2
python-docx
chromadb
sentence-transformers
transformers
torch
